Move anonymize functions outside plugins module
